feat: 게시글 수정 기능 구현

+++ src/main/java/imbesky/simpleforum/controller/ViewerController.java
@@ -27,7 +27,7 @@ public class ViewerController {
     }
 
     @GetMapping("/view")
-    public String postViewer(@RequestParam final int id, final Model model){
+    public String postViewer(@RequestParam final long id, final Model model){
         PostViewDto post = viewerService.inquirePost(id);
         model.addAttribute("post", post);
         return "viewer";
